refactor build supported namespaces
