+++ src/hooks/useProposal/useProposal.ts
@@ -1,19 +1,8 @@
 import { useCallback, useState } from "react";
-import { getErrorMessage, getNamespace } from "@/utils/helper.util";
-import { EIP155_SIGNING_METHODS } from "@/data/methods/EIP155Data.methods";
+import { getErrorMessage } from "@/utils/helper.util";
 import useAnalytics from "@/hooks/useAnalytics";
-import {
-  BIP122_CHAINS,
-  EIP155_CHAINS,
-  MULTIVERS_X_CHAINS,
-  RIPPLE_CHAINS,
-  SupportedNamespace,
-} from "@/data/network.config";
-import {
-  BuildApprovedNamespacesParams,
-  buildApprovedNamespaces,
-} from "@walletconnect/utils";
-import { formatAccountsByChain } from "@/hooks/useProposal/util";
+
+import { buildApprovedNamespaces } from "@walletconnect/utils";
 import { useNavigate } from "@tanstack/react-router";
 import { useQueryClient } from "@tanstack/react-query";
 import {
@@ -31,10 +20,7 @@ import { queryKey as pendingProposalsQueryKey } from "../usePendingProposals";
 import { ProposalTypes } from "@walletconnect/types";
 import { enqueueSnackbar } from "notistack";
 import { sortedRecentConnectionAppsAtom } from "../../store/recentConnectionAppsAtom";
-import { WALLET_METHODS } from "@/data/methods/Wallet.methods";
-import { MULTIVERSX_SIGNING_METHODS } from "@/data/methods/MultiversX.methods";
-import { BIP122_SIGNING_METHODS } from "@/data/methods/BIP122.methods";
-import { RIPPLE_SIGNING_METHODS } from "@/data/methods/Ripple.methods";
+import { useSupportedNamespaces } from "../useSupportedNamespaces";
 
 export function useProposal(proposal: ProposalTypes.Struct) {
   const navigate = useNavigate({ from: "/proposal/$id" });
@@ -44,8 +30,11 @@ export function useProposal(proposal: ProposalTypes.Struct) {
   const walletKit = useAtomValue(walletKitAtom);
   const analytics = useAnalytics();
   const addAppToLastConnectionApps = useSetAtom(sortedRecentConnectionAppsAtom);
-
   const [selectedAccounts, setSelectedAccounts] = useState<string[]>([]);
+  const { buildSupportedNamespaces } = useSupportedNamespaces(
+    proposal,
+    selectedAccounts,
+  );
 
   const handleClick = useCallback((account: string) => {
     setSelectedAccounts((value) => {
@@ -62,295 +51,6 @@ export function useProposal(proposal: ProposalTypes.Struct) {
       search: (search) => search,
     });
   }, [navigate]);
-
-  const buildEip155Namespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        proposal,
-        accounts.data,
-      ).filter((a) => {
-        return (
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(EIP155_CHAINS).includes(a.chain)
-        );
-      });
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(EIP155_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.EIP155]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.EIP155]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.EIP155]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.EIP155]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, proposal, selectedAccounts],
-  );
-
-  const buildMvxNamespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        proposal,
-        accounts.data,
-      ).filter(
-        (a) =>
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(MULTIVERS_X_CHAINS).includes(a.chain),
-      );
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(MULTIVERSX_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.MVX]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.MVX]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.MVX]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.MVX]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, proposal, selectedAccounts],
-  );
-
-  const buildBip122Namespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        proposal,
-        accounts.data,
-      ).filter(
-        (a) =>
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(BIP122_CHAINS).includes(a.chain),
-      );
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(BIP122_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.BIP122]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.BIP122]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.BIP122]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.BIP122]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, proposal, selectedAccounts],
-  );
-
-  const buildXrpNamespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        proposal,
-        accounts.data,
-      ).filter(
-        (a) =>
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(RIPPLE_CHAINS).includes(a.chain),
-      );
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(RIPPLE_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.XRPL]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.XRPL]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.XRPL]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.XRPL]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, proposal, selectedAccounts],
-  );
-
-  const buildSupportedNamespaces = useCallback(
-    (proposal: ProposalTypes.Struct) => {
-      const { requiredNamespaces, optionalNamespaces } = proposal;
-
-      const supportedNamespaces: BuildApprovedNamespacesParams["supportedNamespaces"] =
-        {};
-
-      if ("bip122" in requiredNamespaces || "bip122" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.BIP122] = buildBip122Namespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      if ("eip155" in requiredNamespaces || "eip155" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.EIP155] = buildEip155Namespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      if ("mvx" in requiredNamespaces || "mvx" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.MVX] = buildMvxNamespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      if ("xrpl" in requiredNamespaces || "xrpl" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.XRPL] = buildXrpNamespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      return supportedNamespaces;
-    },
-    [
-      buildBip122Namespace,
-      buildEip155Namespace,
-      buildMvxNamespace,
-      buildXrpNamespace,
-    ],
-  );
 
   const sessionsQueryFn = useSessionsQueryFn(walletKit);
