+++ src/hooks/useSessionDetails.ts
@@ -1,18 +1,5 @@
-import { BIP122_SIGNING_METHODS } from "@/data/methods/BIP122.methods";
-import { EIP155_SIGNING_METHODS } from "@/data/methods/EIP155Data.methods";
-import { MULTIVERSX_SIGNING_METHODS } from "@/data/methods/MultiversX.methods";
-import { RIPPLE_SIGNING_METHODS } from "@/data/methods/Ripple.methods";
-import { WALLET_METHODS } from "@/data/methods/Wallet.methods";
-import {
-  BIP122_CHAINS,
-  EIP155_CHAINS,
-  MULTIVERS_X_CHAINS,
-  RIPPLE_CHAINS,
-  SupportedNamespace,
-} from "@/data/network.config";
 import useAccounts from "@/hooks/useAccounts";
 import useAnalytics from "@/hooks/useAnalytics";
-import { formatAccountsByChain } from "@/hooks/useProposal/util";
 import { walletAPIClientAtom } from "@/store/wallet-api.store";
 import {
   showBackToBrowserModalAtom,
@@ -25,8 +12,7 @@ import {
 } from "@/utils/helper.util";
 import { useQueryClient } from "@tanstack/react-query";
 import { useNavigate } from "@tanstack/react-router";
-import { ProposalTypes, SessionTypes } from "@walletconnect/types";
-import { BuildApprovedNamespacesParams } from "@walletconnect/utils";
+import { SessionTypes } from "@walletconnect/types";
 import { useAtomValue, useSetAtom } from "jotai";
 import { enqueueSnackbar } from "notistack";
 import { useCallback, useEffect, useMemo, useState } from "react";
@@ -36,7 +22,7 @@ import {
   useQueryFn as useSessionsQueryFn,
 } from "./useSessions";
 import { Account } from "@ledgerhq/wallet-api-client";
-import { use } from "i18next";
+import { useSupportedNamespaces } from "./useSupportedNamespaces";
 
 export function useSessionDetails(session: SessionTypes.Struct) {
   const navigate = useNavigate({ from: "/detail/$topic" });
@@ -50,6 +36,11 @@ export function useSessionDetails(session: SessionTypes.Struct) {
   const setShowModal = useSetAtom(showBackToBrowserModalAtom);
   const [updating, setUpdating] = useState(false);
   const [disconnecting, setDisconnecting] = useState(false);
+  const [selectedAccounts, setSelectedAccounts] = useState<string[]>([]);
+  const { buildSupportedNamespaces } = useSupportedNamespaces(
+    session,
+    selectedAccounts,
+  );
 
   const fullAddresses = useMemo(
     () =>
@@ -57,7 +48,7 @@ export function useSessionDetails(session: SessionTypes.Struct) {
         (acc, elem) => acc.concat(elem[1].accounts),
         [] as string[],
       ),
-    [session],
+    [session.namespaces],
   );
 
   const sessionAccounts = useMemo(
@@ -77,16 +68,21 @@ export function useSessionDetails(session: SessionTypes.Struct) {
     }
   }, [sessionAccounts, setMainAccount]);
 
-  const [selectedAccounts, setSelectedAccounts] = useState<string[]>([]);
-
-  const handleClick = useCallback((account: string) => {
-    setSelectedAccounts((value) => {
-      if (value.includes(account)) {
-        return value.filter((s) => s !== account);
-      }
-      return [...value, account];
-    });
-  }, []);
+  const handleClick = useCallback(
+    (account: string) => {
+      setSelectedAccounts((value) => {
+        if (value.includes(account)) {
+          return value.filter((s) => s !== account);
+        }
+        // always put main account first
+        if (mainAccount?.id === account) {
+          return [account, ...value.filter((s) => s !== account)];
+        }
+        return [...value, account];
+      });
+    },
+    [mainAccount?.id],
+  );
 
   const navigateToHome = useCallback(() => {
     return navigate({
@@ -94,295 +90,6 @@ export function useSessionDetails(session: SessionTypes.Struct) {
       search: (search) => search,
     });
   }, [navigate]);
-
-  const buildEip155Namespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        session,
-        accounts.data,
-      ).filter((a) => {
-        return (
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(EIP155_CHAINS).includes(a.chain)
-        );
-      });
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(EIP155_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.EIP155]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.EIP155]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.EIP155]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.EIP155]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, session, selectedAccounts],
-  );
-
-  const buildMvxNamespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        session,
-        accounts.data,
-      ).filter(
-        (a) =>
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(MULTIVERS_X_CHAINS).includes(a.chain),
-      );
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(MULTIVERSX_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.MVX]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.MVX]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.MVX]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.MVX]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, session, selectedAccounts],
-  );
-
-  const buildBip122Namespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        session,
-        accounts.data,
-      ).filter(
-        (a) =>
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(BIP122_CHAINS).includes(a.chain),
-      );
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(BIP122_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.BIP122]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.BIP122]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.BIP122]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.BIP122]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, session, selectedAccounts],
-  );
-
-  const buildXrpNamespace = useCallback(
-    (
-      requiredNamespaces: ProposalTypes.RequiredNamespaces,
-      optionalNamespaces: ProposalTypes.OptionalNamespaces,
-    ) => {
-      const accountsByChain = formatAccountsByChain(
-        session,
-        accounts.data,
-      ).filter(
-        (a) =>
-          a.accounts.length > 0 &&
-          a.isSupported &&
-          Object.keys(RIPPLE_CHAINS).includes(a.chain),
-      );
-      const dataToSend = accountsByChain.reduce<
-        { account: string; chain: string }[]
-      >(
-        (accum, elem) =>
-          accum.concat(
-            elem.accounts
-              .filter((acc) => selectedAccounts.includes(acc.id))
-              .map((a) => ({
-                account: `${getNamespace(a.currency)}:${a.address}`,
-                chain: getNamespace(a.currency),
-              })),
-          ),
-        [],
-      );
-
-      const supportedMethods: string[] = [
-        ...Object.values(WALLET_METHODS),
-        ...Object.values(RIPPLE_SIGNING_METHODS),
-      ];
-
-      const methods = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.XRPL]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-          ...(optionalNamespaces[SupportedNamespace.XRPL]?.methods.filter(
-            (method) => supportedMethods.includes(method),
-          ) ?? []),
-        ]),
-      ];
-
-      const events = [
-        ...new Set([
-          ...(requiredNamespaces[SupportedNamespace.XRPL]?.events ?? []),
-          ...(optionalNamespaces[SupportedNamespace.XRPL]?.events ?? []),
-        ]),
-      ];
-
-      return {
-        chains: [...new Set(dataToSend.map((e) => e.chain))],
-        methods,
-        events,
-        accounts: dataToSend.map((e) => e.account),
-      };
-    },
-    [accounts.data, session, selectedAccounts],
-  );
-
-  const buildSupportedNamespaces = useCallback(
-    (session: SessionTypes.Struct) => {
-      const { requiredNamespaces, optionalNamespaces } = session;
-
-      const supportedNamespaces: BuildApprovedNamespacesParams["supportedNamespaces"] =
-        {};
-
-      if ("bip122" in requiredNamespaces || "bip122" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.BIP122] = buildBip122Namespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      if ("eip155" in requiredNamespaces || "eip155" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.EIP155] = buildEip155Namespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      if ("mvx" in requiredNamespaces || "mvx" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.MVX] = buildMvxNamespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      if ("xrpl" in requiredNamespaces || "xrpl" in optionalNamespaces) {
-        supportedNamespaces[SupportedNamespace.XRPL] = buildXrpNamespace(
-          requiredNamespaces,
-          optionalNamespaces,
-        );
-      }
-      return supportedNamespaces;
-    },
-    [
-      buildBip122Namespace,
-      buildEip155Namespace,
-      buildMvxNamespace,
-      buildXrpNamespace,
-    ],
-  );
 
   const sessionsQueryFn = useSessionsQueryFn(walletKit);
 
@@ -436,7 +143,31 @@ export function useSessionDetails(session: SessionTypes.Struct) {
     }
     setUpdating(true);
     try {
-      const namespaces = buildSupportedNamespaces(session);
+      let namespaces = buildSupportedNamespaces(session);
+
+      // keep the main account in the first position
+      if (mainAccount) {
+        const caipAccount = `${getNamespace(mainAccount.currency)}:${mainAccount?.address}`;
+        const mainAccountNamespace = Object.entries(namespaces).find(
+          ([, value]) => value.accounts.includes(caipAccount),
+        );
+
+        if (mainAccountNamespace) {
+          delete namespaces[mainAccountNamespace[0]];
+          namespaces = {
+            [mainAccountNamespace[0]]: {
+              ...mainAccountNamespace[1],
+              accounts: [
+                caipAccount,
+                ...mainAccountNamespace[1].accounts.filter(
+                  (a) => a !== caipAccount,
+                ),
+              ],
+            },
+            ...namespaces,
+          };
+        }
+      }
 
       await walletKit.updateSession({
         topic: session.topic,
@@ -456,7 +187,7 @@ export function useSessionDetails(session: SessionTypes.Struct) {
       // Remove the uri from the search params to avoid trying to connect again if the user reload the current page
     } catch (error) {
       enqueueSnackbar(getErrorMessage(error), {
-        errorType: "Approve session error",
+        errorType: "Edit session error",
         variant: "errorNotification",
         anchorOrigin: {
           vertical: "top",
@@ -476,22 +207,39 @@ export function useSessionDetails(session: SessionTypes.Struct) {
     selectedAccounts.length,
     buildSupportedNamespaces,
     session,
+    mainAccount,
     walletKit,
     queryClient,
     sessionsQueryFn,
   ]);
 
-  useEffect(() => {
-    console.log("session", session);
-    console.log("selectedAccount", selectedAccounts);
-  }, [session, selectedAccounts]);
-
   const handleSwitch = useCallback(
     async (account: Account) => {
-      if (mainAccount?.id === account.id) return;
-
       const chainId = getNamespace(account.currency);
       const [namespace, chainValue] = chainId.split(":");
+
+      const caipAccount = `${chainId}:${account.address}`;
+      const existingNamespace = session.namespaces[namespace];
+      delete session.namespaces[namespace];
+      session.namespaces = {
+        [namespace]: {
+          ...existingNamespace,
+          accounts: [
+            caipAccount,
+            ...existingNamespace.accounts.filter((a) => a !== caipAccount),
+          ],
+          chains: [
+            chainId,
+            ...(existingNamespace.chains?.filter((c) => c !== chainId) ?? []),
+          ],
+        },
+        ...session.namespaces,
+      };
+
+      await walletKit.updateSession({
+        topic: session.topic,
+        namespaces: session.namespaces,
+      });
 
       if (session.namespaces[namespace].events.includes("chainChanged")) {
         await walletKit.emitSessionEvent({
@@ -528,39 +276,20 @@ export function useSessionDetails(session: SessionTypes.Struct) {
         });
       }
 
-      const caipAccount = `${chainId}:${account.address}`;
-      const existingNamespace = session.namespaces[namespace];
-      delete session.namespaces[namespace];
-
-      await walletKit.updateSession({
-        topic: session.topic,
-        namespaces: {
-          [namespace]: {
-            ...existingNamespace,
-            accounts: [
-              caipAccount,
-              ...existingNamespace.accounts.filter((a) => a !== caipAccount),
-            ],
-            chains: [
-              chainId,
-              ...(existingNamespace.chains?.filter((c) => c !== chainId) ?? []),
-            ],
-          },
-          ...session.namespaces,
-        },
-      });
-
       setMainAccount(account);
       setShowModal(true);
+
+      await queryClient.invalidateQueries({ queryKey: sessionsQueryKey });
+      await queryClient.invalidateQueries({
+        queryKey: pendingProposalsQueryKey,
+      });
+      // Prefetching as we need the data in the next route to avoid redirecting to home
+      await queryClient.prefetchQuery({
+        queryKey: sessionsQueryKey,
+        queryFn: sessionsQueryFn,
+      });
     },
-    [
-      mainAccount?.id,
-      session.namespaces,
-      session.topic,
-      setMainAccount,
-      setShowModal,
-      walletKit,
-    ],
+    [queryClient, session, sessionsQueryFn, setShowModal, walletKit],
   );
 
   function getAccountsFromAddresses(addresses: string[], accounts: Account[]) {
